added additional signature mechanism that needs cleaning, added AddActivityToMany and FollowMany
