--- feed_flat_example_test.go
@@ -1,38 +1,0 @@
-package getstream_test
-
-import (
-	"fmt"
-	"github.com/GetStream/stream-go"
-)
-
-func ExampleFlatFeed_AddActivity() {
-
-	client, err := getstream.New(&getstream.Config{
-		APIKey:    "APIKey",
-		APISecret: "APISecret",
-		AppID:     "AppID",
-		Location:  "Region"})
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	feed, err := client.FlatFeed("FeedSlug", "UserID")
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	activity, err := feed.AddActivity(&getstream.Activity{
-		Verb:      "post",
-		ForeignID: "48d024fe-3752-467a-8489-23febd1dec4e",
-		Object:    getstream.FeedID("flat:eric"),
-		Actor:     getstream.FeedID("flat:john"),
-	})
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	_ = activity
-}
