+++ main.go
@@ -1,0 +1,3 @@
+package getstream
+
+const VERSION = "v1.0.0"
